Change songid to int

+++ src/java/Model/PlaylistSongsDAO.java
@@ -11,13 +11,13 @@ public class PlaylistSongsDAO implements Dao<PlaylistSongs> {
     private DatabaseInformation db = new DatabaseInformation();
 
     @Override
-    public Optional<PlaylistSongs> get(long id) {
+    public Optional<PlaylistSongs> get(int id) {
         PlaylistSongs playlistSongs = null;
-        try (Connection con = db.getConnection()) {
-            try (PreparedStatement stmt = con.prepareStatement(
+        try ( Connection con = db.getConnection()) {
+            try ( PreparedStatement stmt = con.prepareStatement(
                     "SELECT * FROM PlaylistSongs WHERE playlistId = ?")) {
-                stmt.setLong(1, id);
-                try (ResultSet rs = stmt.executeQuery()) {
+                stmt.setInt(1, id);
+                try ( ResultSet rs = stmt.executeQuery()) {
                     if (rs.next()) {
                         playlistSongs = new PlaylistSongs(rs.getInt("playlistId"), rs.getInt("songId"));
                     }
@@ -32,9 +32,7 @@ public class PlaylistSongsDAO implements Dao<PlaylistSongs> {
     @Override
     public ArrayList<PlaylistSongs> getAll() {
         ArrayList<PlaylistSongs> playlistSongsList = new ArrayList<>();
-        try (Connection con = db.getConnection();
-             Statement stmt = con.createStatement();
-             ResultSet rs = stmt.executeQuery("SELECT * FROM PlaylistSongs")) {
+        try ( Connection con = db.getConnection();  Statement stmt = con.createStatement();  ResultSet rs = stmt.executeQuery("SELECT * FROM PlaylistSongs")) {
             while (rs.next()) {
                 PlaylistSongs playlistSongs = new PlaylistSongs(rs.getInt("playlistId"), rs.getInt("songId"));
                 playlistSongsList.add(playlistSongs);
@@ -47,8 +45,8 @@ public class PlaylistSongsDAO implements Dao<PlaylistSongs> {
 
     public boolean save(PlaylistSongs playlistSongs) {
         boolean result = false;
-        try (Connection con = db.getConnection()) {
-            try (PreparedStatement stmt = con.prepareStatement(
+        try ( Connection con = db.getConnection()) {
+            try ( PreparedStatement stmt = con.prepareStatement(
                     "INSERT INTO PlaylistSongs (playlistId, songId) VALUES (?, ?)")) {
                 stmt.setInt(1, playlistSongs.getPlaylistId());
                 stmt.setInt(2, playlistSongs.getSongId());
@@ -63,8 +61,8 @@ public class PlaylistSongsDAO implements Dao<PlaylistSongs> {
     @Override
     public boolean update(PlaylistSongs playlistSongs, String[] params) {
         boolean result = false;
-        try (Connection con = db.getConnection()) {
-            try (PreparedStatement stmt = con.prepareStatement(
+        try ( Connection con = db.getConnection()) {
+            try ( PreparedStatement stmt = con.prepareStatement(
                     "UPDATE PlaylistSongs SET songId = ? WHERE playlistId = ?")) {
                 stmt.setInt(1, Integer.parseInt(params[1]));
                 stmt.setInt(2, Integer.parseInt(params[0]));
@@ -78,8 +76,8 @@ public class PlaylistSongsDAO implements Dao<PlaylistSongs> {
 
     public boolean delete(PlaylistSongs playlistSongs) {
         boolean result = false;
-        try (Connection con = db.getConnection()) {
-            try (PreparedStatement stmt = con.prepareStatement(
+        try ( Connection con = db.getConnection()) {
+            try ( PreparedStatement stmt = con.prepareStatement(
                     "DELETE FROM PlaylistSongs WHERE playlistId = ? AND songId = ?")) {
                 stmt.setInt(1, playlistSongs.getPlaylistId());
                 stmt.setInt(2, playlistSongs.getSongId());
